+++ packages/notifications/push/amplify_push_notifications/android/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'amplify_push_notifications'
