+++ packages/notifications/push/amplify_push_notifications/pubspec.yaml
@@ -11,8 +11,6 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  amplify_push_notifications_ios: ">=1.0.0-next.0 <1.0.0-next.1"
-  amplify_push_notifications_android: ">=1.0.0-next.0 <1.0.0-next.1"
   amplify_secure_storage: ^0.1.0
   amplify_core: ">=1.0.0-next.0 <1.0.0-next.1"
 
@@ -27,6 +25,7 @@ flutter:
   plugin:
     platforms:
       android:
-        default_package: amplify_push_notifications_android
+        package: com.amazonaws.amplify.amplify_push_notifications
+        pluginClass: AmplifyPushNotificationsPlugin
       ios:
-        default_package: amplify_push_notifications_ios
+        pluginClass: AmplifyPushNotificationsPlugin
