--- packages/notifications/push/amplify_push_notifications_android/.metadata
@@ -1,30 +1,0 @@
-# This file tracks properties of this Flutter project.
-# Used by Flutter tool to assess capabilities and perform upgrades etc.
-#
-# This file should be version controlled.
-
-version:
-  revision: a10e295a02d8b3fc2b786ee0f3c1abf602a59956
-  channel: master
-
-project_type: plugin
-
-# Tracks metadata for the flutter migrate command
-migration:
-  platforms:
-    - platform: root
-      create_revision: a10e295a02d8b3fc2b786ee0f3c1abf602a59956
-      base_revision: a10e295a02d8b3fc2b786ee0f3c1abf602a59956
-    - platform: android
-      create_revision: a10e295a02d8b3fc2b786ee0f3c1abf602a59956
-      base_revision: a10e295a02d8b3fc2b786ee0f3c1abf602a59956
-
-  # User provided section
-
-  # List of Local paths (relative to this file) that should be
-  # ignored by the migrate tool.
-  #
-  # Files that are not part of the templates will be ignored by default.
-  unmanaged_files:
-    - 'lib/main.dart'
-    - 'ios/Runner.xcodeproj/project.pbxproj'
