--- packages/notifications/push/amplify_push_notifications_android/analysis_options.yaml
@@ -1,5 +1,0 @@
-include: package:amplify_lints/library.yaml
-
-analyzer:
-  exclude:
-    - "**/*.g.dart"
