--- packages/notifications/push/amplify_push_notifications_android/android/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'amplify_push_notifications_android'
