--- packages/notifications/push/amplify_push_notifications_android/android/src/main/AndroidManifest.xml
@@ -1,21 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.amazonaws.amplify.amplify_push_notifications_android">
-    <application>
-        <service
-            android:name=".PushNotificationFirebaseMessagingService"
-            android:exported="false">
-            <intent-filter>
-                <action android:name="com.google.firebase.MESSAGING_EVENT" />
-            </intent-filter>
-        </service>
-        <receiver
-            android:name=".PushNotificationReceiver"
-            android:exported="true"
-            android:permission="com.google.android.c2dm.permission.SEND">
-
-            <intent-filter>
-                <action android:name="com.google.android.c2dm.intent.RECEIVE" />
-            </intent-filter>
-        </receiver>
-    </application>
-</manifest>
