--- packages/notifications/push/amplify_push_notifications_ios/CHANGELOG.md
@@ -1,3 +1,0 @@
-## 0.0.1
-
-- Initial implementation of amplify_push_notifications_ios package.
