--- packages/notifications/push/amplify_push_notifications_ios/ios/Classes/AmplifyPushNotificationsIosPlugin.swift
@@ -1,9 +1,0 @@
-import Flutter
-import UIKit
-import Amplify
-import amplify_flutter_ios
-
-public class AmplifyPushNotificationsIosPlugin{
-
-}
-
