--- packages/notifications/push/amplify_push_notifications_ios/pubspec.yaml
@@ -1,25 +1,0 @@
-name: amplify_push_notifications_ios
-description: Native iOS Push Notifications Flutter package
-version: 0.0.1
-homepage: https://docs.amplify.aws/lib/q/platform/flutter/
-issue_tracker: https://github.com/aws-amplify/amplify-flutter/issues
-
-environment:
-  sdk: ">=2.17.0 <3.0.0"
-  flutter: ">=3.0.0"
-
-dependencies:
-  flutter:
-    sdk: flutter
-  amplify_core: ">=1.0.0-next.1 <1.0.0-next.2"
-
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  amplify_lints: ^2.0.0
-
-flutter:
-  plugin:
-    platforms:
-      ios:
-        pluginClass: AmplifyPushNotificationsIosPlugin
