+++ packages/notifications/push/amplify_push_notifications_pinpoint/lib/amplify_push_notifications_pinpoint.dart
@@ -4,7 +4,7 @@
 library amplify_push_notifications_pinpoint;
 
 import 'package:amplify_push_notifications/amplify_push_notifications.dart';
-import 'package:amplify_push_notifications_pinpoint/pinpoint_provider.dart';
+import 'package:amplify_push_notifications_pinpoint/src/pinpoint_provider.dart';
 
 /// {@template amplify_push_notifications_pinpoint.amplify_push_notifications_pinpoint}
 /// The AWS Pinpoint implementation of the Amplify Push Notification category.
